피드백 1차 반영
